test: fold the suites onto the option each one covers
Twenty-four files for seventeen subjects: `generate` was spread over three,
`minify` over three, `minimizerOptions` over two, and `imagemin` over two.
Each option now has one file -- watch, the filesystem cache and
`imageminGenerate` join `generate`, css and swc-html join `minify`,
`terserOptions` joins the alias it is deprecated for, and
`imageminNormalizeConfig` joins the minimizer that calls it. The helpers
those files each carried a copy of are now written once.

Merging needed the environment gate to move first. Four of them were skipped
whole by `testPathIgnorePatterns`, so folding one into a file that runs
everywhere would have taken it to Node 10 and to Windows, where its
minimizers do not install. `test/helpers/env.js` now holds what this
environment can run as one answer: `jest.config.js` reads it to skip a whole
file, and a merged file reads it to skip only the block that needs more.
Two files still need the whole-file kind, both because they reach for what
they need at module scope.

`getEcmaVersion` is covered by builds now rather than by calling it: each
case sets `output.environment` and reads back the `ecma` that reached the
minimizer, which is the wiring the option is for. Every flag is stated false
first, since webpack fills an unset one from the target and would raise the
answer behind the case; and the minimizer runs in process, or the recording
happens in a worker and never comes back.

Snapshots moved with their tests and were re-keyed for the describe level the
gating adds. Same 573 tests, same 825 snapshots.
