+++ test/generate-option.test.js
@@ -1,10 +1,20 @@
+import fs from "fs";
+import os from "os";
 import path from "path";
 
 import MinimizerPlugin from "../src";
 import { replaceExtension } from "../src/utils";
 
 import { compile, getCompiler, getErrors, getWarnings } from "./helpers";
+import { RUN_IMAGE_TESTS } from "./helpers/env";
 
+/**
+ * `describe` where this environment can run the block, `describe.skip` where it
+ * cannot, so one file can carry blocks with different requirements.
+ * @param {boolean} condition whether this environment can run it
+ * @returns {jest.Describe} describe, or describe.skip
+ */
+const describeIf = (condition) => (condition ? describe : describe.skip);
 // Renaming an asset needs `NormalModule`'s `processResult` hook to be able to
 // await. Read off what the build did rather than off a version number: the
 // release carrying it is not out yet, so a version test would claim the
@@ -34,6 +44,8 @@ const IMAGE_RULES = [
 function toWebp(input) {
   const [[name, code]] = Object.entries(input);
 
+  toWebp.calls += 1;
+
   return {
     code: Buffer.concat([Buffer.from("WEBP:"), Buffer.from(code)]),
     filename: replaceExtension(name, "webp"),
@@ -42,6 +54,29 @@ function toWebp(input) {
 
 toWebp.supportsBinary = () => true;
 toWebp.supportsWorker = () => false;
+toWebp.calls = 0;
+
+/**
+ * A second encoder. Its source differs from `toWebp`'s deliberately: source is
+ * what tells two generators apart when neither reports a version.
+ * @param {{ [file: string]: string | Buffer }} input input
+ * @returns {{ code: Buffer, filename: string }} the re-encoded result
+ */
+function toAvif(input) {
+  const [[name, code]] = Object.entries(input);
+  const marker = Buffer.from("AVIF:");
+
+  toAvif.calls += 1;
+
+  return {
+    code: Buffer.concat([marker, Buffer.from(code)]),
+    filename: replaceExtension(name, "avif"),
+  };
+}
+
+toAvif.supportsBinary = () => true;
+toAvif.supportsWorker = () => false;
+toAvif.calls = 0;
 
 /**
  * @param {import("webpack").Compiler} compiler compiler
@@ -53,6 +88,104 @@ function readBytes(compiler, stats, name) {
   return compiler.outputFileSystem.readFileSync(
     path.join(stats.compilation.outputOptions.path, name),
   );
+}
+
+/**
+ * @param {string} directory directory to remove, with everything under it
+ * @returns {void}
+ */
+function removeRecursive(directory) {
+  for (const entry of fs.readdirSync(directory)) {
+    const full = path.join(directory, entry);
+
+    if (fs.statSync(full).isDirectory()) {
+      removeRecursive(full);
+    } else {
+      fs.unlinkSync(full);
+    }
+  }
+
+  fs.rmdirSync(directory);
+}
+
+/**
+ * @param {import("webpack").Stats} stats stats
+ * @param {string} name the module's short name
+ * @returns {{ cached: boolean, built: boolean }} how it came to be in the build
+ */
+function moduleState(stats, name) {
+  const { modules } = stats.toJson({
+    all: false,
+    modules: true,
+    cachedModules: true,
+  });
+  const found = modules.find((item) => item.name === name);
+
+  return { cached: Boolean(found.cached), built: Boolean(found.built) };
+}
+
+/**
+ * Drives a watching compiler one build at a time: each call resolves with the
+ * stats of the next build the watcher completes.
+ */
+class Watcher {
+  /**
+   * @param {import("webpack").Compiler} compiler compiler
+   */
+  constructor(compiler) {
+    this.pending = [];
+    this.waiting = [];
+    // Polling, because the CI runners disagree about native file watching.
+    this.watching = compiler.watch(
+      { aggregateTimeout: 50, poll: 100 },
+      (error, stats) => {
+        const settle = this.waiting.shift();
+
+        if (settle) {
+          settle(error, stats);
+        } else {
+          this.pending.push([error, stats]);
+        }
+      },
+    );
+  }
+
+  /**
+   * @returns {Promise<import("webpack").Stats>} the stats of the next build
+   */
+  next() {
+    return new Promise((resolve, reject) => {
+      /**
+       * @param {(Error | null)=} error build error
+       * @param {import("webpack").Stats=} stats build stats
+       * @returns {void}
+       */
+      const settle = (error, stats) => {
+        if (error) {
+          reject(error);
+        } else {
+          resolve(/** @type {import("webpack").Stats} */ (stats));
+        }
+      };
+
+      const ready = this.pending.shift();
+
+      if (ready) {
+        settle(ready[0], ready[1]);
+      } else {
+        this.waiting.push(settle);
+      }
+    });
+  }
+
+  /**
+   * @returns {Promise<void>} resolves once the watcher has let go of the files
+   */
+  close() {
+    return new Promise((resolve) => {
+      this.watching.close(() => resolve());
+    });
+  }
 }
 
 describe("generate option", () => {
@@ -281,6 +414,405 @@ describe("sharpGenerate target format", () => {
 
     expect(result.errors).toHaveLength(1);
     expect(String(result.errors[0])).toMatch(/does not write 'bmp'/);
+  });
+});
+
+describe("imageminGenerate", () => {
+  // The only block here needing the image packages, so it is gated on its own
+  // rather than the whole file being skipped where they are absent.
+  describeIf(RUN_IMAGE_TESTS)("with `imagemin` installed", () => {
+    // SVG markup under a name claiming a raster format: the same mismatch
+    // `imageminMinify` refuses, which is the one a generator exists to take.
+    const svg = Buffer.from(
+      '<svg xmlns="http://www.w3.org/2000/svg"><rect x="1.00000"/></svg>',
+    );
+
+    it("should rename an image a plugin turned into SVG", async () => {
+      const { code, filename, warnings } =
+        await MinimizerPlugin.imageminGenerate(
+          { "photo.png": svg },
+          undefined,
+          { plugins: ["svgo"] },
+        );
+
+      expect(warnings).toBeUndefined();
+      expect(filename).toBe("photo.svg");
+      // svgo ran: the padded coordinate is what it trims.
+      expect(code.toString()).not.toContain("1.00000");
+      expect(code.toString()).toContain("<svg");
+    });
+
+    it("should keep the name when the format did not change", async () => {
+      const { code, filename } = await MinimizerPlugin.imageminGenerate(
+        { "photo.svg": svg },
+        undefined,
+        { plugins: ["svgo"] },
+      );
+
+      expect(filename).toBeUndefined();
+      expect(code.toString()).not.toContain("1.00000");
+    });
+
+    it("should keep the query and fragment the name carried", async () => {
+      const { filename } = await MinimizerPlugin.imageminGenerate(
+        { "photo.png?w=100#frag": svg },
+        undefined,
+        { plugins: ["svgo"] },
+      );
+
+      expect(filename).toBe("photo.svg?w=100#frag");
+    });
+
+    it("should declare what it needs from the plugin", () => {
+      expect(MinimizerPlugin.imageminGenerate.supportsBinary()).toBe(true);
+      // Its plugins shell out to native binaries, so it cannot leave the process.
+      expect(MinimizerPlugin.imageminGenerate.supportsWorker()).toBe(false);
+      expect(MinimizerPlugin.imageminGenerate.supportsWorkerThreads()).toBe(
+        false,
+      );
+      expect(MinimizerPlugin.imageminGenerate.filter("photo.png")).toBe(true);
+      expect(MinimizerPlugin.imageminGenerate.filter("main.js")).toBe(false);
+    });
+  });
+});
+
+describe("generate option in watch mode", () => {
+  let context;
+  let watcher;
+
+  beforeEach(() => {
+    toWebp.calls = 0;
+    context = fs.mkdtempSync(path.join(os.tmpdir(), "minimizer-watch-"));
+
+    fs.writeFileSync(
+      path.join(context, "index.js"),
+      'import jpg from "./image.jpg";\n\n// eslint-disable-next-line no-console\nconsole.log(jpg);\n',
+    );
+    fs.writeFileSync(path.join(context, "image.jpg"), Buffer.from("first"));
+  });
+
+  afterEach(async () => {
+    if (watcher) {
+      await watcher.close();
+      watcher = undefined;
+    }
+
+    removeRecursive(context);
+  });
+
+  /**
+   * @returns {import("webpack").Compiler} a compiler over the temporary project
+   */
+  function makeCompiler() {
+    const compiler = getCompiler({
+      context,
+      entry: path.join(context, "index.js"),
+      // `production` leaves caching off, and then every watch build rebuilds
+      // every module — which is not what a rename has to survive.
+      cache: { type: "memory" },
+      module: {
+        rules: [
+          {
+            test: /\.jpe?g$/i,
+            type: "asset/resource",
+            generator: { filename: "[name][ext]" },
+          },
+        ],
+      },
+    });
+
+    new MinimizerPlugin({ test: /\.jpe?g$/i, generate: toWebp }).apply(
+      compiler,
+    );
+
+    return compiler;
+  }
+
+  /**
+   * @param {import("webpack").Compiler} compiler compiler
+   * @param {import("webpack").Stats} stats stats
+   * @param {string} name emitted name
+   * @returns {Buffer} the emitted bytes
+   */
+  function readBytes(compiler, stats, name) {
+    return compiler.outputFileSystem.readFileSync(
+      path.join(stats.compilation.outputOptions.path, name),
+    );
+  }
+
+  it("should re-emit the renamed asset when the image changes", async () => {
+    const compiler = makeCompiler();
+
+    watcher = new Watcher(compiler);
+
+    const first = await watcher.next();
+
+    if (reportedNoAwait(first)) {
+      return;
+    }
+
+    expect(getErrors(first)).toEqual([]);
+    expect(Object.keys(first.compilation.assets)).toContain("image.webp");
+    expect(readBytes(compiler, first, "image.webp").toString()).toBe(
+      "WEBP:first",
+    );
+
+    fs.writeFileSync(path.join(context, "image.jpg"), Buffer.from("second"));
+
+    const second = await watcher.next();
+
+    expect(getErrors(second)).toEqual([]);
+
+    const names = Object.keys(second.compilation.assets);
+
+    expect(names).toContain("image.webp");
+    expect(names).not.toContain("image.jpg");
+    // The generator's answer is cached on the bytes, so new bytes have to
+    // reach it rather than the first build's result being served again.
+    expect(toWebp.calls).toBe(2);
+    expect(readBytes(compiler, second, "image.webp").toString()).toBe(
+      "WEBP:second",
+    );
+  });
+
+  it("should keep the rename when only the module importing it changes", async () => {
+    const compiler = makeCompiler();
+
+    watcher = new Watcher(compiler);
+
+    const first = await watcher.next();
+
+    if (reportedNoAwait(first)) {
+      return;
+    }
+
+    expect(Object.keys(first.compilation.assets)).toContain("image.webp");
+
+    fs.writeFileSync(
+      path.join(context, "index.js"),
+      'import jpg from "./image.jpg";\n\n// eslint-disable-next-line no-console\nconsole.log(jpg, "changed");\n',
+    );
+
+    const second = await watcher.next();
+
+    expect(getErrors(second)).toEqual([]);
+
+    const names = Object.keys(second.compilation.assets);
+
+    // One call across both builds is the evidence that the image module was
+    // not rebuilt, so the rename survived on the module rather than being
+    // reapplied — `buildInfo.assetResource` is what carries it.
+    expect(toWebp.calls).toBe(1);
+    expect(names).toContain("image.webp");
+    expect(names).not.toContain("image.jpg");
+
+    const bundle = readBytes(compiler, second, "main.js").toString();
+
+    expect(bundle).toContain('"image.webp"');
+    expect(bundle).not.toContain('"image.jpg"');
+  });
+});
+
+describe("generate option with the filesystem cache", () => {
+  let context;
+  let cacheDirectory;
+
+  beforeEach(() => {
+    toWebp.calls = 0;
+    toAvif.calls = 0;
+    context = fs.mkdtempSync(path.join(os.tmpdir(), "minimizer-fs-cache-"));
+    cacheDirectory = path.join(context, "cache");
+
+    fs.writeFileSync(
+      path.join(context, "index.js"),
+      'import jpg from "./image.jpg";\n\n// eslint-disable-next-line no-console\nconsole.log(jpg);\n',
+    );
+    fs.writeFileSync(path.join(context, "image.jpg"), Buffer.from("first"));
+  });
+
+  afterEach(() => {
+    removeRecursive(context);
+  });
+
+  /**
+   * Runs one build against the shared cache directory and closes the compiler,
+   * which is what writes the pack out for the next run to read.
+   * @param {object=} options plugin options overriding the defaults
+   * @returns {Promise<{ stats: import("webpack").Stats, assets: string[], read: (name: string) => string }>} what the build produced
+   */
+  function run(options) {
+    const compiler = getCompiler({
+      context,
+      entry: path.join(context, "index.js"),
+      // Through the config rather than a later `apply`: the cache strategy is
+      // built while webpack applies the configured plugins, so a plugin
+      // applied after `webpack()` returns cannot reach its version.
+      plugins: [
+        new MinimizerPlugin({
+          test: /\.jpe?g$/i,
+          generate: toWebp,
+          ...options,
+        }),
+      ],
+      cache: {
+        type: "filesystem",
+        cacheDirectory,
+        // The test writes no config file for webpack to watch, and the
+        // default points at one that does not exist here.
+        buildDependencies: {},
+      },
+      module: {
+        rules: [
+          {
+            test: /\.jpe?g$/i,
+            type: "asset/resource",
+            generator: { filename: "[name][ext]" },
+          },
+        ],
+      },
+    });
+
+    return new Promise((resolve, reject) => {
+      compiler.run((error, stats) => {
+        if (error) {
+          compiler.close(() => reject(error));
+
+          return;
+        }
+
+        // Nothing is read here: on a webpack that cannot await, the build
+        // reports an error and emits no bundle, and the caller checks that
+        // before asking for one.
+        const assets = Object.keys(stats.compilation.assets);
+        const output = stats.compilation.outputOptions.path;
+
+        compiler.close((closeError) => {
+          if (closeError) {
+            reject(closeError);
+
+            return;
+          }
+
+          resolve({
+            stats,
+            assets,
+            read: (name) =>
+              compiler.outputFileSystem
+                .readFileSync(path.join(output, name))
+                .toString(),
+          });
+        });
+      });
+    });
+  }
+
+  it("should keep the rename when the module is restored from the pack", async () => {
+    const first = await run();
+
+    if (reportedNoAwait(first.stats)) {
+      return;
+    }
+
+    expect(getErrors(first.stats)).toEqual([]);
+    expect(first.assets).toContain("image.webp");
+    expect(toWebp.calls).toBe(1);
+
+    const second = await run();
+
+    expect(getErrors(second.stats)).toEqual([]);
+
+    // The second compiler is a new one reading the pack the first wrote, so
+    // the image module is restored rather than rebuilt. `assetResource` is
+    // serialized with it, which is what has to carry the rename -- a
+    // `matchResource` would not survive.
+    expect(toWebp.calls).toBe(1);
+    expect(moduleState(second.stats, "./image.jpg")).toEqual({
+      cached: true,
+      built: false,
+    });
+    expect(second.assets).toContain("image.webp");
+    expect(second.assets).not.toContain("image.jpg");
+    const bundle = second.read("main.js");
+
+    expect(bundle).toContain('"image.webp"');
+    expect(bundle).not.toContain('"image.jpg"');
+  });
+
+  it("should re-encode when the image changed between runs", async () => {
+    const first = await run();
+
+    if (reportedNoAwait(first.stats)) {
+      return;
+    }
+
+    expect(first.assets).toContain("image.webp");
+
+    fs.writeFileSync(path.join(context, "image.jpg"), Buffer.from("second"));
+
+    const second = await run();
+
+    expect(getErrors(second.stats)).toEqual([]);
+    // New bytes, so neither the module nor the generator's own cache entry
+    // may answer from the pack.
+    expect(moduleState(second.stats, "./image.jpg").built).toBe(true);
+    expect(toWebp.calls).toBe(2);
+    expect(second.assets).toContain("image.webp");
+    expect(second.assets).not.toContain("image.jpg");
+  });
+
+  it("should re-run a changed generator against a warm pack", async () => {
+    const first = await run();
+
+    if (reportedNoAwait(first.stats)) {
+      return;
+    }
+
+    expect(first.assets).toContain("image.webp");
+
+    const second = await run({ generate: toAvif });
+
+    expect(getErrors(second.stats)).toEqual([]);
+    // Nothing per-module keys on the plugin, so without the generator in the
+    // pack's version the restored module would keep the previous generator's
+    // bytes and name.
+    expect(toAvif.calls).toBe(1);
+    expect(second.assets).toContain("image.avif");
+    expect(second.assets).not.toContain("image.webp");
+    expect(second.read("image.avif")).toBe("AVIF:first");
+  });
+
+  it("should re-run the generator when only its options changed", async () => {
+    const first = await run({ generatorOptions: { quality: 50 } });
+
+    if (reportedNoAwait(first.stats)) {
+      return;
+    }
+
+    expect(toWebp.calls).toBe(1);
+
+    const second = await run({ generatorOptions: { quality: 90 } });
+
+    expect(getErrors(second.stats)).toEqual([]);
+    expect(toWebp.calls).toBe(2);
+    expect(second.assets).toContain("image.webp");
+  });
+
+  it("should read an array of generators for the identity too", async () => {
+    const first = await run({ generate: [toWebp] });
+
+    if (reportedNoAwait(first.stats)) {
+      return;
+    }
+
+    expect(first.assets).toContain("image.webp");
+    expect(toWebp.calls).toBe(1);
+
+    const second = await run({ generate: [toAvif] });
+
+    expect(getErrors(second.stats)).toEqual([]);
+    expect(toAvif.calls).toBe(1);
+    expect(second.assets).toContain("image.avif");
+    expect(second.assets).not.toContain("image.webp");
   });
 });
